feat: auth.ts
